refactor(ui): extract shared PipelineSchematicBase to eliminate SVG duplication
All 9 pipeline schematic components were near-identical copies (~175 lines
each). Extract the shared rendering logic into a single PipelineSchematicBase
component that accepts row0, row1, ariaLabel, and optional extraColors props.
Each schematic is now a ~15-line thin wrapper that declares only its step
data arrays. Reduces duplicated code by ~80% (~1400 → ~285 lines).
